Rename scripts/generators to .cjs for ESM migration
