+++ example/rosidl/rosidl-parse-action-example.js
@@ -15,7 +15,7 @@
 const rosInstallPath = process.env.AMENT_PREFIX_PATH;
 const packageName = 'test_msgs';
 const packagePath = rosInstallPath + '/share/test_msgs/action/Fibonacci.action';
-const parser = require('../../rosidl_parser/rosidl_parser.js');
+const parser = require('../../rosidl_parser/rosidl_parser.cjs');
 
 parser
   .parseActionFile(packageName, packagePath)
