+++ example/rosidl/rosidl-parse-srv-example.js
@@ -14,7 +14,7 @@
 
 'use strict';
 
-const parser = require('../../rosidl_parser/rosidl_parser.js');
+const parser = require('../../rosidl_parser/rosidl_parser.cjs');
 
 const rosInstallPath = process.env.AMENT_PREFIX_PATH;
 const packageName = 'std_srvs';
