+++ lib/interface_loader.js
@@ -16,7 +16,7 @@
 
 const path = require('path');
 const fs = require('fs');
-const generator = require('../rosidl_gen/index.js');
+const generator = require('../rosidl_gen/index.cjs');
 const { TypeValidationError, ValidationError } = require('./errors.js');
 
 let interfaceLoader = {
