+++ rosidl_gen/index.cjs
@@ -15,10 +15,10 @@
 'use strict';
 
 const fse = require('../lib/utils.js');
-const generateJSStructFromIDL = require('./idl_generator.js');
-const packages = require('./packages.js');
+const generateJSStructFromIDL = require('./idl_generator.cjs');
+const packages = require('./packages.cjs');
 const path = require('path');
-const idlConvertor = require('../rosidl_convertor/idl_convertor.js');
+const idlConvertor = require('../rosidl_convertor/idl_convertor.cjs');
 const generatedRoot = path.join(__dirname, '../generated/');
 const serviceMsgPath = path.join(generatedRoot, 'srv_msg');
 const idlPath = path.join(generatedRoot, 'share');
@@ -58,7 +58,7 @@ function generateInPathSyncWorker(targetPath) {
     // Use child_process.spawnSync for truly synchronous execution
     const result = require('child_process').spawnSync(
       'node',
-      [path.join(__dirname, 'generate_worker.js')],
+      [path.join(__dirname, 'generate_worker.cjs')],
       {
         env: { ...process.env, WORKER_TARGET_PATH: targetPath },
         encoding: 'utf8',
