+++ test/test-array-data.js
@@ -17,7 +17,7 @@
 const assert = require('assert');
 const deepEqual = require('deep-equal');
 const rclnodejs = require('../index.js');
-const translator = require('../rosidl_gen/message_translator.js');
+const translator = require('../rosidl_gen/message_translator.cjs');
 const arrayGen = require('./array_generator.js');
 const { isTypedArray } = require('./utils.js');
